homepage_service/requirements.txt
fastapi
uvicorn
requests
python-multipart
prometheus_fastapi_instrumentator
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http